% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate.R
\name{get_acceleration}
\alias{get_acceleration}
\alias{put_acceleration}
\title{Bucket Acceleration}
\usage{
get_acceleration(bucket, ...)

put_acceleration(bucket, status = c("Enabled", "Suspended"), ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}

\item{status}{Character string specifying whether acceleration should be \dQuote{Enabled} or \dQuote{Suspended}.}
}
\value{
For \code{get_acceleration}: If acceleration has never been enabled or suspend, the value is \code{NULL}. Otherwise, the status is returned (either \dQuote{Enabled} or \dQuote{Suspended}). For \code{put_acceleration}: If acceleration has never been enabled or suspend, the value is \code{NULL}.
' @examples
\dontrun{
b <- bucketlist()
get_acceleration(b[[1]])
put_acceleration(b[[1]], "Enabled")
get_acceleration(b[[1]])
put_acceleration(b[[1]], "Suspended")
}
}
\description{
Get/Put acceleration settings or retrieve acceleration status of a bucket.
}
