% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{bucket_exists}
\alias{bucket_exists}
\title{Bucket exists?}
\usage{
bucket_exists(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
\code{TRUE} if bucket exists and is accessible, else \code{FALSE}. An \code{qingstor_error} object may be returned if the request failed.
}
\description{
Check whether a bucket exists and is accessible with the current authentication keys.
}
