% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketlist.R
\name{bucketlist}
\alias{bucketlist}
\title{List Buckets}
\usage{
bucketlist(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
a list of buckets.  if passed with default settings \dQuote{parse_response = TRUE}
the response will list the name and creationdate of all buckets owned by the request
sender.
}
\description{
List buckets for authenticated user
}
\details{
\code{bucketlist} performs a GET operation on the base Qing endpoint 
and returns a list of all buckets owned by the authenticated 
sender of the request. If authentication is successful, this function provides a list of 
buckets available to the authenticated user. In this way, it can serve as 
a \dQuote{hello world!} function, to confirm that one's authentication 
credentials are working correctly.
}
\keyword{service}
