% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cors.R
\name{get_cors}
\alias{get_cors}
\alias{put_cors}
\alias{delete_cors}
\title{CORS}
\usage{
get_cors(bucket, ...)

put_cors(bucket, ...)

delete_cors(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
For \code{get_cors}: A list with cors configuration and rules. For \code{delete_cors}: \code{TRUE} if successful, \code{FALSE} otherwise.
}
\description{
Get/Put/Delete the cross origin resource sharing configuration information for a bucket.
}
