% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete Bucket}
\usage{
delete_bucket(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise. 
An \code{qingstor_error} object may be returned if the request failed.
}
\description{
Deletes an Qing bucket.
}
