% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT.R
\name{delete_object}
\alias{delete_object}
\title{Delete object}
\usage{
delete_object(object, bucket, quiet = TRUE, ...)
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{object}. In most cases, if \code{object} is specified as the latter.}

\item{bucket}{Character string with the name of the bucket.}
}
\value{
\code{TRUE} if successful, otherwise an object of class Qingstor_error details if not.
}
\description{
Deletes one objects from an Qingstor bucket.
}
\examples{
# delete an object in Qingstor bucket

delete_object(object = 'object_name', bucket = 'bucket_name')
the object_name could be a name under a folder like 'some_folder/object_name'

}
\seealso{
\code{\link{put_object}}
}
