% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_bucket}
\alias{get_bucket}
\title{Get bucket}
\usage{
get_bucket(bucket, prefix = NULL, delimiter = NULL, max = NULL,
  marker = NULL, parse_response = TRUE, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{prefix}{Character string that limits the response to keys that begin 
with the specified prefix}

\item{parse_response}{logical, should we attempt to parse the response?}

\item{folder}{must, now set to "/"}
}
\value{
A dataFrame of objects infomation in the bucket.
}
\description{
List the contents of an Qingstor bucket
}
\examples{
# list the object in Qingstor bucket

get_bucket(bucket = 'bucket_name',folder = "/")
prefix is support ,example:
get_bucket(bucket = 'bucket_name',folder = "/",prefix = 'prefix_name')

}
