% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_location}
\alias{get_location}
\title{Bucket location}
\usage{
get_location(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
A character string containing the region, if one has been set.
}
\description{
Get the qingstor region location of bucket.
}
