% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getobject.R
\name{get_torrent}
\alias{get_torrent}
\title{Get object torrent}
\usage{
get_torrent(object, bucket, ...)
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{object}. In most cases, if \code{object} is specified as the latter.}

\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
Something.
}
\description{
Retrieves a Bencoded dictionary (BitTorrent) for an object from an Qing bucket.
}
