% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_uploads}
\alias{get_uploads}
\title{Multipart uploads}
\usage{
get_uploads(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
A list containing the multipart upload information.
}
\description{
Get a list of multipart uploads for a bucket.
}
