% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{get_bucket_policy}
\alias{get_bucket_policy}
\alias{put_bucket_policy}
\alias{delete_bucket_policy}
\title{Bucket policies}
\usage{
get_bucket_policy(bucket, ...)

put_bucket_policy(bucket, policy, ...)

delete_bucket_policy(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}

\item{policy}{A character string containing a bucket policy.}
}
\value{
For \code{get_policy}: A character string containing the JSON representation of the policy, if one has been set. For \code{delete_policy} and \code{put_policy}: \code{TRUE} if successful, \code{FALSE} otherwise. An \code{qingstor_error} object may be returned if the request failed.
}
\description{
Get/Put/Delete the bucket access policy for a bucket.
}
