% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{put_bucket}
\alias{put_bucket}
\title{Create bucket}
\usage{
put_bucket(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
\code{TRUE} if successful, qingstor_error details if not.
}
\description{
Creates a new Qing bucket.
}
