% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT.R
\name{put_object}
\alias{put_object}
\title{Put object}
\usage{
put_object(file, object, bucket, headers = list(), ...)
}
\arguments{
\item{file}{A character string containing the filename (or full path) of 
the file you want to upload to Qingstor.}

\item{object}{A character string containing the name the object should 
have in Qingstor (i.e., its "object key"). If missing, the filename is used.}

\item{bucket}{Character string with the name of the bucket.}

\item{headers}{List of request headers for the REST call.}

\item{folder}{A character string specify the folder name in the bucket. Optional}
}
\value{
If successful, \code{TRUE}, otherwise an qingstor_error object.
}
\description{
Puts an object into an Qingstor bucket
}
\examples{
# put an object in Qingstor bucket
tempPath <- file.path("path","to","folder","xxxx.csv")
put_object(file = tempPath , object = basename(tempPath),folder = 'folder_name', bucket = 'bucket_name')

}
\seealso{
\code{\link{delete_object}}
}
