% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qingcurl.R
\name{qingcurl}
\alias{qingcurl}
\title{qingcurl}
\usage{
qingcurl(object, bucket, open = "rb", accelerate = FALSE,
  region = Sys.getenv("Qstor_DEFAULT_REGION", "pek3a"),
  key = Sys.getenv("Qstor_ACCESS_KEY_ID"),
  secret = Sys.getenv("Qstor_SECRET_ACCESS_KEY"))
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{object}. In most cases, if \code{object} is specified as the latter.}

\item{bucket}{Character string with the name of the bucket.}

\item{open}{One of \dQuote{r} or \dQuote{rb}.}

\item{region}{A character string containing the QingStor region. Ignored if region can be inferred from \code{bucket}.
If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an Qingstor Access Key ID. 
If missing, defaults to value stored in environment variable \dQuote{Qstor_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an QingStor Secret Access Key. 
If missing, defaults to value stored in environment variable \dQuote{Qstor_SECRET_ACCESS_KEY}.}
}
\value{
If \code{file = NULL}, a raw object. Otherwise, a character string containing the file name that the object is saved to.
}
\description{
Create an qingcurl connection
}
\details{
This function creates a \code{\link[curl]{curl}} connection to be used for asynchronous reading from an Qing object.
}
\seealso{
\code{\link{get_object}}, \code{\link{save_object}}, \code{\link[curl]{curl}}
}
