% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QingHTTP.R
\name{QingHTTP}
\alias{QingHTTP}
\title{Qing HTTP Requests}
\usage{
QingHTTP(verb = "GET", bucket = "", path = "", folder = "",
  query = NULL, headers = list(), request_body = "", accelerate = FALSE,
  region = Sys.getenv("Qstor_DEFAULT_REGION", "pek3a"),
  key = Sys.getenv("Qstor_ACCESS_KEY_ID"),
  secret = Sys.getenv("Qstor_SECRET_ACCESS_KEY"),
  session_token = Sys.getenv("Qstor_SESSION_TOKEN"), parse_response = TRUE,
  ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{bucket}{A character string with the name of the bucket, or an object of class \dQuote{qing_bucket}. If the latter and a region can be inferred from the bucket object attributes, then that region is used instead of \code{region}.}

\item{path}{A character string with the name of the object to put in the bucket 
(sometimes called the object or 'key name' in the QingStor documentation.)}

\item{query}{any queries, passed as a named list}

\item{headers}{a list of request headers for the REST call.}

\item{request_body}{character string of request body data.}

\item{accelerate}{A logical indicating whether to use QingStor transfer acceleration, which can produce significant speed improvements for cross-country transfers. Acceleration only works with buckets that do not have dots in bucket name.}

\item{region}{A character string containing the QingStor region. Ignored if region can be inferred from \code{bucket}. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an QingStor Access Key ID. If missing, defaults to value stored in environment variable \dQuote{Qstor_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an QingStor Secret Access Key. If missing, defaults to value stored in environment variable \dQuote{Qstor_SECRET_ACCESS_KEY}.}

\item{parse_response}{return the response as is, or parse and return as a list? Default is TRUE.}

\item{...}{Additional arguments passed to an HTTP request function. such as \code{\link[httr]{GET}}.}
}
\value{
the Qing response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for Qing.
}
\details{
This is mostly an internal function for executing API requests. In almost all cases, users do not need to access this directly.
}
