% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qingsave.R
\name{Qingsave}
\alias{Qingsave}
\alias{Qingsave_image}
\alias{Qingload}
\title{save/load}
\usage{
Qingsave(..., object, bucket, opts = NULL)

Qingsave_image(object, bucket, opts = NULL)

Qingload(object, bucket, envir = parent.frame(), ...)
}
\arguments{
\item{...}{For \code{Qingsave}, one or more R objects to be saved via \code{\link[base]{save}} and uploaded to Qing. For \code{Qingload}, see \code{opts}.}

\item{object}{For \code{Qingsave}, a character string of the name of the object you want to save to. For \code{Qingload}, a character string of the name of the object you want to load from Qing.}

\item{bucket}{Character string with the name of the bucket.}

\item{opts}{Additional arguments passed to \code{\link{QingHTTP}}.}

\item{envir}{An R environment to load objects into. Default is the \code{parent.frame()} from which the function is called.}
}
\value{
For \code{Qingsave}, a logical, invisibly, otherwise an error object. For \code{Qingload}, \code{NULL} invisibly, otherwise an error object.
}
\description{
Save/load R object(s) to/from Qing
}
\examples{
\dontrun{
# create bucket
b <- put_bucket("myexamplebucket")

# save a dataset to the bucket
Qingsave(mtcars, iris, object = "somedata.Rdata", bucket = b)
get_bucket(b)

# load the data from bucket
e <- new.env()
Qingload(object = "somedata.Rdata", bucket = b, envir = e)
ls(e)

# cleanup
rm(e)
delete_object(object = "somedata.Rdata", bucket = "myexamplebucket")
delete_bucket("myexamplebucket")
}
}
\seealso{
\code{\link{QingsaveRDS}},\code{\link{QingreadRDS}}
}
