% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QingsaveRDS.R
\name{QingsaveRDS}
\alias{QingsaveRDS}
\alias{QingreadRDS}
\title{saveRDS/readRDS}
\usage{
QingsaveRDS(x, bucket, object = paste0(as.character(substitute(x)), ".rds"),
  ...)

QingreadRDS(bucket, object, ...)
}
\arguments{
\item{x}{For \code{QingsaveRDS}, a single R object to be saved via \code{\link[base]{saveRDS}} and uploaded to Qing. \code{x} is analogous to the \code{object} argument in \code{saveRDS}.}

\item{bucket}{Character string with the name of the bucket.}

\item{object}{Character string with the object key, or an object of class \dQuote{object}. In most cases, if \code{object} is specified as the latter.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
For \code{QingsaveRDS}, a logical. For \code{QingreadRDS}, an R object.
}
\description{
Serialization interface to read/write R objects to Qing
}
\examples{
\dontrun{
# create bucket
b <- put_bucket("myexamplebucket")

# save a single object to Qing
QingsaveRDS(x = mtcars, bucket = "myexamplebucket", object = "mtcars.rds")

# restore it under a different name
mtcars2 <- QingreadRDS(object = "mtcars.rds", bucket = "myexamplebucket")
identical(mtcars, mtcars2)

# cleanup
delete_object(object = "mtcars.rds", bucket = "myexamplebucket")
delete_bucket("myexamplebucket")
}
}
\seealso{
\code{\link{Qingsave}},\code{\link{Qingload}}
}
\author{
Steven Akins <skawesome@gmail.com>
}
