% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qingsource.R
\name{Qingsource}
\alias{Qingsource}
\title{Source from Qing}
\usage{
Qingsource(object, bucket, ..., opts = NULL)
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{object}. In most cases, if \code{object} is specified as the latter.}

\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}

\item{opts}{Additional arguments passed to \code{\link{get_object}} for retrieving the R syntax file.}
}
\value{
See \code{\link[base]{source}}
}
\description{
Source R code (a la \code{\link[base]{source}}) from Qing
}
\examples{
\dontrun{
# create bucket
b <- put_bucket("myexamplebucket")

# save some code to the bucket
cat("x <- 'hello world!'\\nx", file = "example.R")
put_object("example.R", object = "example.R", bucket = b)
get_bucket(b)

# source the code from the bucket
Qingsource(object = "example.R", bucket = b, echo = TRUE)

# cleanup
unlink("example.R")
delete_object(object = "example.R", bucket = b)
delete_bucket("myexamplebucket")
}
}
\seealso{
\code{\link{QingsaveRDS}},\code{\link{Qingsave}}, \code{\link{get_object}}
}
