% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qingstor-package.R
\docType{package}
\name{Qingstor-package}
\alias{Qingstor-package}
\alias{Qingstor}
\title{Qingstor-package}
\description{
Qingstor R SDK
}
\details{
First Set Sys.setenv("Qstor_ACCESS_KEY_ID" = "", "Qstor_SECRET_ACCESS_KEY" = "", "Qstor_DEFAULT_REGION" = "")
}
\author{
Xiao Ma
}
\keyword{package}
