% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestPayment.R
\name{get_requestpayment}
\alias{get_requestpayment}
\alias{put_requestpayment}
\title{requestPayment}
\usage{
get_requestpayment(bucket, ...)

put_requestpayment(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{\dots}{Additional arguments passed to \code{\link{QingHTTP}}.}
}
\value{
A list containing the requestPayment information, if set.
}
\description{
Get/Put the requestPayment subresource for a bucket.
}
